requirement.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==2.2
django-environ==0.4.5
djangorestframework==3.12.4
geojson==2.5.0
idna==2.10
ipython==5.0.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
mock==4.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pyowm==2.10.0
pytz==2021.1
requests==2.25.1
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
urllib3==1.26.4
wcwidth==0.2.5
